+++ contrib/scripts/readall.sh
@@ -11,7 +11,7 @@ if [ "x$1" = "x-R" ]; then
   readargs=$1
   shift
 fi
-for i in `echo "f -F circuit,name" "$@"|nc -q 1 127.0.0.1 $port|sort -u|egrep ','`; do
+for i in `echo "f -F circuit,name" "$@"|nc -q 1 127.0.0.1 $port|sort -u|grep ','`; do
   circuit=${i%%,*}
   name=${i##*,}
   if [ -z "$circuit" ] || [ -z "$name" ] || [ "$circuit,$name" = "scan,id" ]; then