+++ src/lib/ebus/datatype.cpp
@@ -1,6 +1,6 @@
 /*
  * ebusd - daemon for communication with eBUS heating systems.
- * Copyright (C) 2014-2023 John Baier <ebusd@ebusd.eu>
+ * Copyright (C) 2014-2024 John Baier <ebusd@ebusd.eu>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -591,7 +591,8 @@ result_t DateTimeDataType::writeSymbols(size_t offset, size_t length, istringstr
         if (result != RESULT_OK) {
           return result;  // invalid time part
         }
-        if ((i == 0 && value > 24) || (i > 0 && (last == 24 && value > 0) )) {
+        if ((i == (m_hasDate ? 2 : 0) && value > 24)
+        || (i > (m_hasDate ? 2 : 0) && (last == 24 && value > 0) )) {
           return RESULT_ERR_OUT_OF_RANGE;  // invalid time part
         }
         if (hasFlag(SPE)) {  // minutes since midnight