+++ ChangeLog.md
@@ -1,3 +1,12 @@
+# next (tbd)
+## Bug Fixes
+* fix conditional messages not being sent to message definition in MQTT integration and not being used in KNX group association
+* fix CSV dump of config files on command line
+* fix DTM type with recent dates
+## Features
+* add "inject" command
+
+
 # 23.3 (2023-12-26)
 ## Bug Fixes
 * fix MQTT topic string validation